change location of manual add
